+++ crankshaft-console/src/input.rs
@@ -16,10 +16,7 @@ pub fn should_ignore_key_event(input: &Event) -> bool {
 pub fn should_quit(input: &Event) -> bool {
     use Event::*;
     use KeyCode::*;
-    match input {
-        Key(KeyEvent {
-            code: Char('q'), ..
-        }) => true,
+    matches!(input,
         Key(KeyEvent {
             code: Char('c'),
             modifiers,
@@ -29,7 +26,49 @@ pub fn should_quit(input: &Event) -> bool {
             code: Char('d'),
             modifiers,
             ..
-        }) if modifiers.contains(KeyModifiers::CONTROL) => true,
-        _ => false,
-    }
+        }) if modifiers.contains(KeyModifiers::CONTROL))
+}
+
+/// Returns true if the user wants to go to the next task.
+pub fn is_next_task(input: &Event) -> bool {
+    matches!(
+        input,
+        Event::Key(KeyEvent {
+            code: KeyCode::Char('j') | KeyCode::Down,
+            ..
+        })
+    )
+}
+
+/// Returns true if the user wants to go to the previous task.
+pub fn is_previous_task(input: &Event) -> bool {
+    matches!(
+        input,
+        Event::Key(KeyEvent {
+            code: KeyCode::Char('k') | KeyCode::Up,
+            ..
+        })
+    )
+}
+
+/// Returns true if the user wants to see the task view.
+pub fn is_view_tasks(input: &Event) -> bool {
+    matches!(
+        input,
+        Event::Key(KeyEvent {
+            code: KeyCode::Char('t'),
+            ..
+        })
+    )
+}
+
+/// Returns true if the user wants to cancel the selected task.
+pub fn is_cancel_task(input: &Event) -> bool {
+    matches!(
+        input,
+        Event::Key(KeyEvent {
+            code: KeyCode::Char('c'),
+            ..
+        })
+    )
 }