+++ crankshaft-engine/src/service/runner/backend/docker.rs
@@ -376,6 +376,8 @@ impl crate::Backend for Backend {
         let events = self.events.clone();
         let names = self.names.clone();
 
+        let task_token = CancellationToken::new();
+
         Ok(async move {
             // Generate a name of the task if one wasn't provided
             let task_name = task.name.unwrap_or_else(|| {
@@ -483,7 +485,9 @@ impl crate::Backend for Backend {
                         select! {
                             // Always poll the cancellation token first
                             biased;
-
+                            _= task_token.cancelled() => {
+                                (Err(TaskRunError::Canceled), Cleanup::Service(service))
+                            }
                             _ = token.cancelled() => {
                                 (Err(TaskRunError::Canceled), Cleanup::Service(service))
                             }
@@ -530,7 +534,9 @@ impl crate::Backend for Backend {
                         select! {
                             // Always poll the cancellation token first
                             biased;
-
+                            _ = task_token.cancelled() => {
+                                (Err(TaskRunError::Canceled), Cleanup::Container(container))
+                            }
                             _ = token.cancelled() => {
                                 (Err(TaskRunError::Canceled), Cleanup::Container(container))
                             }
@@ -541,7 +547,8 @@ impl crate::Backend for Backend {
                     };
 
                     if run_cleanup {
-                        cleanup.run(token.is_cancelled()).await?;
+                        let force_remove = matches!(result, Err(TaskRunError::Canceled));
+                        cleanup.run(force_remove).await?;
                     }
 
                     outputs.push(result?);
@@ -553,7 +560,7 @@ impl crate::Backend for Backend {
             };
 
             // Send the created event
-            send_event!(events, Event::TaskCreated { id: task_id, name: task_name.clone(), tes_id: None });
+            send_event!(events, Event::TaskCreated { id: task_id, name: task_name.clone(), tes_id: None, token:task_token.clone()  });
 
             // Run the task to completion
             let result = run.await;