+++ crankshaft-monitor/src/service.rs
@@ -18,6 +18,8 @@ use tonic::Result;
 use tonic::Status;
 use tracing::error;
 
+use crate::proto::CancelTaskRequest;
+use crate::proto::CancelTaskResponse;
 use crate::proto::Event;
 use crate::proto::ExitStatus;
 use crate::proto::ServiceStateRequest;
@@ -76,9 +78,9 @@ impl IntoProtobuf<ExitStatus> for std::process::ExitStatus {
 impl IntoProtobuf<EventKind> for CrankshaftEvent {
     fn into_protobuf(self) -> EventKind {
         match self {
-            CrankshaftEvent::TaskCreated { id, name, tes_id } => {
-                EventKind::Created(TaskCreatedEvent { id, name, tes_id })
-            }
+            CrankshaftEvent::TaskCreated {
+                id, name, tes_id, ..
+            } => EventKind::Created(TaskCreatedEvent { id, name, tes_id }),
             CrankshaftEvent::TaskStarted { id } => EventKind::Started(TaskStartedEvent { id }),
             CrankshaftEvent::TaskContainerCreated { id, container } => {
                 EventKind::ContainerCreated(TaskContainerCreatedEvent { id, container })
@@ -134,6 +136,8 @@ impl IntoProtobuf<Event> for CrankshaftEvent {
 pub struct ServiceState {
     /// The map of task identifier to its events.
     tasks: HashMap<TaskId, TaskEvents>,
+    /// The map of task identifier to its CancellationToken
+    tokens: HashMap<TaskId, CancellationToken>,
 }
 
 /// Represents a gRPC service for monitoring Crankshaft events in real-time.
@@ -173,8 +177,8 @@ impl MonitorService {
                 r = events.recv() => match r {
                     Ok(event) => {
                         let (id, remove) = match event {
-                            CrankshaftEvent::TaskCreated { id, .. }
-                            | CrankshaftEvent::TaskStarted { id }
+                            CrankshaftEvent::TaskCreated { id, .. } |
+                            CrankshaftEvent::TaskStarted { id }
                             | CrankshaftEvent::TaskContainerCreated { id, .. }
                             | CrankshaftEvent::TaskContainerExited { id, .. }
                             | CrankshaftEvent::TaskStdout { id, .. }
@@ -188,17 +192,17 @@ impl MonitorService {
                         if remove {
                             let mut state = state.write().await;
                             state.tasks.remove(&id);
+                            state.tokens.remove(&id);
+                        } else if let CrankshaftEvent::TaskCreated { token, ..} = &event {
+                            let token = token.clone();
+                            let event: Event = event.into_protobuf();
+                            let mut state = state.write().await;
+                            state.tasks.insert(id, TaskEvents { events: vec![event] });
+                            state.tokens.insert(id, token);
                         } else {
                             let event: Event = event.into_protobuf();
                             let mut state = state.write().await;
-                            let task = state.tasks.entry(id).or_default();
-
-                            // If there aren't any events, ensure the first one is the created event
-                            if task.events.is_empty() {
-                                if let Some(EventKind::Created(_)) = &event.event_kind {
-                                    task.events.push(event);
-                                }
-                            } else {
+                            if let Some(task) = state.tasks.get_mut(&id) {
                                 task.events.push(event);
                             }
                         }
@@ -249,5 +253,21 @@ impl Monitor for MonitorService {
         Ok(Response::new(ServiceStateResponse {
             tasks: state.tasks.clone(),
         }))
+    }
+
+    async fn cancel_task(
+        &self,
+        request: Request<CancelTaskRequest>,
+    ) -> Result<Response<CancelTaskResponse>, Status> {
+        let id = request.into_inner().id;
+
+        let state = self.state.read().await;
+
+        if let Some(token) = state.tokens.get(&id) {
+            token.cancel();
+            Ok(Response::new(CancelTaskResponse {}))
+        } else {
+            Err(Status::not_found(format!("Task `{id}` not found")))
+        }
     }
 }