+++ crankshaft-monitor/tests/server.rs
@@ -7,6 +7,7 @@ use std::process::ExitStatus;
 
 use crankshaft_events::Event as CrankshaftEvent;
 use crankshaft_monitor::Monitor;
+use crankshaft_monitor::proto::CancelTaskRequest;
 use crankshaft_monitor::proto::ServiceStateRequest;
 use crankshaft_monitor::proto::SubscribeEventsRequest;
 use crankshaft_monitor::proto::event::EventKind;
@@ -19,6 +20,7 @@ use tokio_retry2::Retry;
 use tokio_retry2::RetryError;
 use tokio_retry2::strategy::ExponentialFactorBackoff;
 use tokio_retry2::strategy::MaxInterval;
+use tokio_util::sync::CancellationToken;
 
 #[tokio::test]
 async fn test_subscribe_events() {
@@ -51,6 +53,7 @@ async fn test_subscribe_events() {
         id: 0,
         name: "first".into(),
         tes_id: None,
+        token: CancellationToken::new(),
     })
     .unwrap();
     tx.send(CrankshaftEvent::TaskStarted { id: 0 }).unwrap();
@@ -63,6 +66,7 @@ async fn test_subscribe_events() {
         id: 1,
         name: "second".into(),
         tes_id: Some("tes".into()),
+        token: CancellationToken::new(),
     })
     .unwrap();
     tx.send(CrankshaftEvent::TaskStarted { id: 1 }).unwrap();
@@ -154,6 +158,7 @@ async fn test_service_state() {
         id: 0,
         name: "first".into(),
         tes_id: None,
+        token: CancellationToken::new(),
     })
     .unwrap();
     tx.send(CrankshaftEvent::TaskStarted { id: 0 }).unwrap();
@@ -166,6 +171,7 @@ async fn test_service_state() {
         id: 1,
         name: "second".into(),
         tes_id: Some("tes".into()),
+        token: CancellationToken::new(),
     })
     .unwrap();
     tx.send(CrankshaftEvent::TaskStarted { id: 1 }).unwrap();
@@ -211,6 +217,46 @@ async fn test_service_state() {
         }
         _ => panic!("unexpected event"),
     }
+
+    drop(tx);
+    monitor.stop().await;
+}
+
+#[tokio::test]
+async fn test_cancel_task() {
+    let (tx, _) = broadcast::channel(16);
+
+    let monitor = Monitor::start("127.0.0.1:32002".parse().unwrap(), tx.clone());
+
+    let strategy = ExponentialFactorBackoff::from_millis(50, 2.0)
+        .max_interval(1000)
+        .take(10);
+
+    let mut client = Retry::spawn(strategy, || async {
+        MonitorClient::connect("http://127.0.0.1:32002")
+            .await
+            .map_err(RetryError::transient)
+    })
+    .await
+    .expect("failed to connect to monitor");
+
+    let token = CancellationToken::new();
+
+    // Send some dummy events at the start
+    tx.send(CrankshaftEvent::TaskCreated {
+        id: 0,
+        name: "first".into(),
+        tes_id: None,
+        token: token.clone(),
+    })
+    .unwrap();
+
+    client
+        .cancel_task(CancelTaskRequest { id: 0 })
+        .await
+        .expect("failed to cancel task");
+
+    token.cancelled().await;
 
     drop(tx);
     monitor.stop().await;